+++ common/src/main/java/io/github/mortuusars/chalk/world/chalk/MarkDrawingContext.java
@@ -1,7 +1,7 @@
 package io.github.mortuusars.chalk.world.chalk;
 
-import io.github.mortuusars.chalk.util.Codecs;
 import io.github.mortuusars.chalk.util.GridCell;
+import io.github.mortuusars.mortaar.network.codec.StreamCodecs;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.network.FriendlyByteBuf;
@@ -11,8 +11,8 @@ import net.minecraft.world.phys.Vec3;
 
 public record MarkDrawingContext(InteractionHand hand, Vec3 clickLocation, BlockPos markPos, Direction markFacing) {
     public static final StreamCodec<FriendlyByteBuf, MarkDrawingContext> STREAM_CODEC = StreamCodec.composite(
-          Codecs.INTERACTION_HAND, MarkDrawingContext::hand,
-          Codecs.VEC3, MarkDrawingContext::clickLocation,
+          StreamCodecs.INTERACTION_HAND, MarkDrawingContext::hand,
+          StreamCodecs.VEC3, MarkDrawingContext::clickLocation,
           BlockPos.STREAM_CODEC, MarkDrawingContext::markPos,
           Direction.STREAM_CODEC, MarkDrawingContext::markFacing,
           MarkDrawingContext::new